+++ packages/granularity/src/components/GrCheckboxGroup/config.ts
@@ -1,0 +1,6 @@
+import { defineGranularComponent } from '@feugene/unocss-preset-granular/contract'
+
+export const grCheckboxGroupConfig = defineGranularComponent(import.meta.url, {
+  name: 'GrCheckboxGroup',
+  dependencies: ['GrCheckbox'],
+})